lumien-backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
alembic>=1.13.1
cors>=1.0.1
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
reportlab>=4.0.9
email-validator>=2.1.0.post1
pycryptodome>=3.20.0
